+++ CHANGELOG.md
@@ -36,6 +36,8 @@ and this project adheres to [PEP 440 Versioning](https://peps.python.org/pep-044
   - Pipeline does not produce auto-generated `requirements.txt` anymore. <!-- TG-2459 -->
 - **Pipeline Raw App**:
   - Now offers the ability to create simple project skeleton. <!-- TG-2495 -->
+- **Configuration**:
+  - Default project configuration now includes the YouWol's project templates. <!-- TG-2497 -->
 
 ### Fixed